Move `goe_embedding` to `wiring`
